fix(ui): guard agent detail time labels against Invalid Date
formatTime lacked the finite TimeClip check used by formatDate and
formatRelativeShort, so malformed createdAt values could still render
as the native Invalid Date string beside a correct date fallback.

Extract pure formatters and cover malformed, out-of-range, null, and
valid inputs.

Closes #18812.
